Simplify the WrappedClusterManager
