+++ vertx-core/src/test/java/io/vertx/tests/eventbus/ClusteredEventBusTest.java
@@ -45,7 +45,7 @@ import java.util.stream.Stream;
 public class ClusteredEventBusTest extends ClusteredEventBusTestBase {
 
   @Test
-  public void testLocalHandlerNotVisibleRemotely() throws Exception {
+  public void testLocalHandlerNotVisibleRemotely() {
     startNodes(2);
     vertices[1].eventBus().localConsumer(ADDRESS1).handler(msg -> {
       fail("Should not receive message");
@@ -396,21 +396,15 @@ public class ClusteredEventBusTest extends ClusteredEventBusTestBase {
     CountDownLatch updateLatch = new CountDownLatch(3);
     startNodes(2, () -> new WrappedClusterManager(getClusterManager()) {
       @Override
-      public void registrationListener(RegistrationListener registrationListener) {
-        super.registrationListener(new WrappedNodeSelector((NodeSelector) registrationListener) {
-          @Override
-          public void registrationsUpdated(RegistrationUpdateEvent event) {
-            super.registrationsUpdated(event);
-            if (event.address().equals(ADDRESS1) && event.registrations().size() == 1) {
-              updateLatch.countDown();
-            }
-          }
-
-          @Override
-          public boolean wantsUpdatesFor(String address) {
-            return true;
-          }
-        });
+      public void registrationsUpdated(RegistrationUpdateEvent event) {
+        super.registrationsUpdated(event);
+        if (event.address().equals(ADDRESS1) && event.registrations().size() == 1) {
+          updateLatch.countDown();
+        }
+      }
+      @Override
+      public boolean wantsUpdatesFor(String address) {
+        return true;
       }
     });
     waitFor(2);
@@ -486,17 +480,10 @@ public class ClusteredEventBusTest extends ClusteredEventBusTestBase {
 
   @Test
   public void testSelectorWantsUpdates() {
-    AtomicReference<NodeSelector> nodeSelectorRef = new AtomicReference<>();
-    startNodes(1, () -> new WrappedClusterManager(getClusterManager()) {
-      @Override
-      public void registrationListener(RegistrationListener registrationListener) {
-        nodeSelectorRef.set((NodeSelector) registrationListener);
-        super.registrationListener(registrationListener);
-      }
-    });
-    assertNotNull(nodeSelectorRef.get());
+    WrappedClusterManager wrapped = new WrappedClusterManager(getClusterManager());
+    startNodes(1, () -> wrapped);
     vertices[0].eventBus().consumer(ADDRESS1, msg -> {
-      assertTrue(nodeSelectorRef.get().wantsUpdatesFor(ADDRESS1));
+      assertTrue(wrapped.wantsUpdatesFor(ADDRESS1));
       testComplete();
     }).completion().onComplete(onSuccess(v -> vertices[0].eventBus().send(ADDRESS1, "foo")));
     await();
@@ -504,15 +491,9 @@ public class ClusteredEventBusTest extends ClusteredEventBusTestBase {
 
   @Test
   public void testSelectorDoesNotWantUpdates() {
-    AtomicReference<NodeSelector> nodeSelectorRef = new AtomicReference<>();
-    startNodes(1, () -> new WrappedClusterManager(getClusterManager()) {
-      @Override
-      public void registrationListener(RegistrationListener registrationListener) {
-        nodeSelectorRef.set((NodeSelector) registrationListener);
-      }
-    });
-    assertNotNull(nodeSelectorRef.get());
-    assertFalse(nodeSelectorRef.get().wantsUpdatesFor(ADDRESS1));
+    WrappedClusterManager wrapped = new WrappedClusterManager(getClusterManager());
+    startNodes(1, () -> wrapped);
+    assertFalse(wrapped.wantsUpdatesFor(ADDRESS1));
   }
 
   @Test
